Remove `express_cache` feature flag (#1145)
## Description of change

- Removes the feature flag so the shared cache may be included in the
next build;
- Adds a changelog entry introducing the feature.

(update and merge this after:
https://github.com/awslabs/mountpoint-s3/pull/1144)

Relevant issues: N/A

## Does this change impact existing behavior?

No, a new feature added.

## Does this change need a changelog entry in any of the crates?

Yes, adding one in this PR.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>