Update documentation to fix installation on Ubuntu 24.04 (#1150)
On Ubuntu 24.04, installation was failing with the following error:
"mount-s3 : Depends: libfuse2 but it is not installable".

This change tells users to update the package index, such that the
needed package `libfuse2t64` can be found, fixing installation errors in
Ubuntu 24.04.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@hagemeier.ch>
Signed-off-by: Christian Hagemeier <chagem@amazon.com>