Fix warnings for test struct variant not used (#1151)
This addresses the only build warning we have in Mountpoint's own
crates. The remaining build warnings come from the fuser forked crate,
which we plan to address through an upstream contribution.

### Does this change impact existing behavior?

No, avoids import of unused code in a test only.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>