Refactor MemoryLimiter to specify tracked 'area' when reserving memory (#1161)
The memory limiter currently tracks the amount of memory reserved for
prefetching. We plan to extend this as part of supporting appends in S3
Express One Zone (#1160).

This change (originally authored by @monthonk) refactors the memory
limiter API to allow specifying the "area" we'd like to reserve in, for
the purpose of metrics for now.

### Does this change impact existing behavior?

No change to existing behavior.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>
Co-authored-by: Monthon Klongklaew <monthonk@amazon.com>