Improve support for additional checksum algorithms in mountpoint-s3-client (#1157)
Allows to specify any of the supported checksum algorithms when
uploading objects.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry?

Yes, adding an entry to the `mountpoint-s3-client` changelog.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>