Upgrade cargo-deny action, remove deprecated deny configurations (#1168)
The `cargo-deny` action we depend on release v2 in August, and with it
made breaking changes. Dependabot was unable to merge due to these
breaking changes: https://github.com/awslabs/mountpoint-s3/pull/969/

This change removes the deprecated configurations. `cargo-deny` now
marks all of those we configured as denied rather than allowing the
violations to be downgraded to warnings or allowed. This impacts us only
for 'unmaintained' crates which is fine, if needed we can always create
an exception entry.

### Does this change impact existing behavior?

This is a CI change only. We upgrade, removing unused and deprecated
fields. Unmaintained crates will now fail CI.

### Does this change need a changelog entry?

No, there is no customer-facing change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>