Fix flaky write_with_sse_kms_key_id_ok test (#1140)
Fixes the flakiness of write_with_sse_kms_key_id_ok test, which was
caused by not properly unmounting and dropping child.

This is not a breaking change; no changelog entry required (as this just
fixes a test).


Before this change, this test fails in ~10 out of 100 runs, after this
change it fails 0 times out of 100 runs.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Christian Hagemeier <chagem@amazon.com>