Fix flaky out_of_order_write test (#1170)
The `out_of_order_write` tests failed to account that in incremental
upload mode the previously written content of a file could be already
uploaded when a subsequent `write` fails (as expected in the test). In
this case, the upload would occasionally be triggered by a `flush` call
as a consequence of the test runner process being forked.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>