Update get_object_attributes test to account for default checksum (#1185)
The [improved support for object integrity
checks](https://docs.aws.amazon.com/AmazonS3/latest/userguide/checking-object-integrity.html#using-additional-checksums)
that is being rolled out in Amazon S3 means that GetObjectAttributes
returns a `CRC-64NVME` object checksum even when no checksum was
specified when the object was uploaded.

This change removes the expectation of an empty checksum that was
causing a test failure in some regions.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>