Refactor incremental upload queue (#1181)
Internal refactor of the append queue for incremental uploads. Splits up
the initial `HeadObject` request and return the checksum algorithm of
the existing object separately from the `PutObject` responses.

### Does this change impact existing behavior?

No, internal change only.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>