Simplify get_object by waiting for response headers (#1171)
`S3CrtClient::get_object` was originally implemented so that it would
complete immediately and return a `GetObjectRequest` implementation
(extending `Stream`) to retrieve body parts. Any error from the S3
request would be returned through the stream.
We recently added additional methods (`get_object_metadata` in #1065 and
`get_object_checksum` in #1123) to the response that rely on the headers
returned by the (first) `GetObject` request. The new methods required an
async signature and a complicated implementation in order to account for
failures and they still do not correctly report accurate error
information in some cases.
With this change, we modify `get_object` to await for response headers
before returning either an error or a `GetObjectResponse` (note the name
change) implementation. The ergonomics of `get_object` are improved:
* `await`ing the initial call can already return some errors (e.g.
bucket/key not found),
* `get_object_checksum` and `get_object_metadata` are now sync
functions.

### Does this change impact existing behavior?

Yes, `get_object` behavior is different, `get_object_checksum` and
`get_object_metadata` signatures have changed, and `GetObjectRequest`
was renamed to `GetObjectResponse`.

### Does this change need a changelog entry?

Yes, it requires a breaking change entry for `mountpoint-s3-client`.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>