Allow partial repeat of `readdir` response (#965)
## Description of change
When user application gets interrupted in a `readdir` syscall the
underlying chain of `readdir` fuse requests gets reset to an offset
which is considered stale by Mountpoint. In that case Mountpoint still
completes the interrupted `readdir` request, but kernel partially
discards the response. We already cache the last response, so we can use
it to serve the request which follows the interrupt.

Relevant issues: https://github.com/awslabs/mountpoint-s3/issues/955

## Does this change impact existing behavior?

This is not a breaking change. Previously an error was returned, now
it'll be handled properly.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vladislav Volodkin <vlaad@amazon.co.uk>
Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vladislav Volodkin <vlaad@amazon.co.uk>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>