Add label to errno in error message for clarity (#1197)
This change improves the clarity of [fuse_error with
errono](https://github.com/awslabs/mountpoint-s3/pull/1189).

### Does this change impact existing behavior? No

### Does this change need a changelog entry? No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Renan Magagnin <renanmagagnin@gmail.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>