Create new latency benchmark for writing of one-byte files (#1190)
Creates a new latency benchmark for writing of one-byte files. Also
creates a folder for writing latency benchmarks and extends the
`fs_latency_bench.sh` script to handle multiple folders.

### Does this change impact existing behavior? No

### Does this change need a changelog entry? No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Renan Magagnin <renanmagagnin@gmail.com>