Improve get_object interface for backpressure
Currently, we support flow-control window for GetObject requests by allowing
applications to call `GetObjectResponse::increment_read_window` but it is
tricky to use because we need to hold onto the stream itself in order to
control the feedback loop while also consuming the data.

This change introduces a new trait `ClientBackpressureHandle` for controlling
the read window so that the stream and the flow-control paths are decoupled.

Applications can now call `GetObjectResponse::take_backpressure_handle`
to get a backpressure handle from the response and use this handle to
extend the read window.

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>