Remove workaround for macOS builds due to pkg-config issues (#1202)
In November, macOS runners had an issue with the `pkg-config` package
and we introduced a workaround:
https://github.com/awslabs/mountpoint-s3/pull/1158/

We see now that the fix has been released and we should no longer need
the workaround: https://github.com/actions/runner-images/issues/10984.

### Does this change impact existing behavior?

No, CI change only.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>