Improve get_object interface for backpressure (#1200)
Currently, we support flow-control window for GetObject requests by
allowing applications to call `GetObjectResponse::increment_read_window`
but it is tricky to use because we need to hold onto the stream itself
in order to control the feedback loop while also consuming the data.

This change introduces a new trait `ClientBackpressureHandle` for
controlling the read window so that the stream and the flow-control
paths are decoupled.

Applications can now call `GetObjectResponse::take_backpressure_handle`
to get a backpressure handle from the response and use this handle to
extend the read window.

### Does this change impact existing behavior?

Yes, there is a breaking change for `mountpoint-s3-client`.

### Does this change need a changelog entry?

Yes, for `mountpoint-s3-client`.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Monthon Klongklaew <monthonk@amazon.com>