Address shadowing divergence in reftest, update semantics doc (#1201)
This commit addresses a case where MP model and property tests diverge
(https://github.com/awslabs/mountpoint-s3/pull/1066). The issue was
caused by the reference not correctly implementing the shadowing order
defined in
[#4f8cf0b](https://github.com/awslabs/mountpoint-s3/commit/4f8cf0b7054d2ea4dedb11ce28c6847849d2eb53).
This commit fixes the reference model, and clarifies the semantics
arising from concurrent MPUs.

This is not a breaking change, as it only impacts the reference tests.

This does not need a Changelog entry, as the change does not impact
Mountpoint's behaviour.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Christian Hagemeier <chagem@amazon.com>