Store the name of a mounted bucket in block's metadata (#1208)
The field `x-amz-meta-source-bucket-name` of the cache block was
intended to store the name of the mounted bucket (source bucket).
Currently it stores the name of the cache bucket.

### Does this change impact existing behavior?

Yes, we update the version of the block schema. All blocks written with
previous versions of Mountpoint won't be accessible (attempts will be
cache misses).

### Does this change need a changelog entry?

Yes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Signed-off-by: Volodkin Vladislav <vladvolodkin@gmail.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>