Improve tracing and assertion messages for reftests (#1211)
This change adds some additional tracing to reftests and makes some
adjustments to assertion messages to make it clearer why we assert what
we assert and would return a better message when things go wrong.

There are no significant changes, this is primarily readability and
debugging improvements.

### Does this change impact existing behavior?

No change to behavior of Mountpoint or its libraries.

### Does this change need a changelog entry?

No, no behavior change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>