Allow clients to define custom callbacks to handle telemetry (#1080)
## Description of change
Different users of mountpoint will care about different metrics returned
for each requests, so allow them to define their own custom handlers for
the on_telemetry callback in addition to the default metrics that
mountpoint emits.

This allows users to do things like: 
- emit extended request ids ("x-amz-id-2")
- When some criteria is met, log out additional information

Relevant issues: #1079 

## Does this change impact existing behavior?


No there should be no breaking changes, the only visible change is that
there's a new field to the S3ClientConfig which defines the custom
telemetry handler

## Does this change need a changelog entry in any of the crates?

Just a note in mountpoint-s3-client letting users know this feature now
exists

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Hans Pistor <hpistor@amazon.com>
Signed-off-by: aws-hans-pistor <133711035+aws-hans-pistor@users.noreply.github.com>
Co-authored-by: Volodkin Vladislav <vladvolodkin@gmail.com>