Create file-system-benchmarks.rs and update fs_latency_bench.sh accordingly (#1213)
Creation of a benchmarking binary for recreating and measuring the
latencies of real-world usage patterns of Mountpoint. This version
includes only a small file creation benchmark which measures the
sequence of the file system operations: lookup, open, write (of one
byte), and flush. The latency measurement captures the total duration
and is averaged multiple iterations to ensure representativeness.

The benchmarking binary is used in the `fs_latency_bench.sh` script and
the the new results are included in the final results of the script,
ultimately being added to the [benchmarking GitHub pages
dashboard](https://awslabs.github.io/mountpoint-s3/dev/latency_bench/).

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Renan Magagnin <renanmagagnin@gmail.com>