Add detailed rustdoc to reftests (#1232)
The reference tests for Mountpoint can be quite complex, especially for
those unfamiliar both with the tests themselves or the idea of
reference-based testing.

This change adds more detailed rustdoc with the goal to ramp up new
readers with the reftests, give an overview of what the tests are doing,
and point the reader to resources for learning more.

### Does this change impact existing behavior?

No, all code documentation changes.

### Does this change need a changelog entry? Does it require a version
change?

No, code doc changes only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>