Add source uri to headers for COPY request (#1228)
This changes is to address gap in supporting buckets with dots in the
name for COPY requests.
First encountered in s3-torch-connector
https://github.com/awslabs/s3-connector-for-pytorch/issues/295

### Does this change impact existing behavior?

No

### Does this change need a changelog entry?

Yes

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Ilya Isaev <iisaev@amazon.com>
Signed-off-by: Isaev Ilya <iisaev@amazon.co.uk>
Co-authored-by: Ilya Isaev <iisaev@amazon.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>