Publish client crates (#1237)
Update changelogs for the client crates to prepare for publication. 

### Does this change impact existing behavior?

N/A

### Does this change need a changelog entry? Does it require a version
change?

N/A

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>