Update CRT submodules to latest releases (#1242)
Update the CRT to the latest releases.

This change also updates the exclude list, primarily due to one of the
test files being replaced by a compressed (but still large) file:
https://github.com/aws/aws-lc/pull/2123/

This change pulls in a bug fix
(https://github.com/awslabs/aws-c-auth/pull/259/), addressing
https://github.com/awslabs/mountpoint-s3/issues/1207.

### Does this change impact existing behavior?

One bug fix is included in CRT changes.

### Does this change need a changelog entry? Does it require a version
change?

Change log entry added for the CRT fix. It is a bug fix, so patch
version bump to `mountpoint-s3-client` remains appropriate.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>