Update S3ClientConfig to support configurable EventLoop thread count (#1240)
For our use case, we run many mountpoint-s3 clients on a single machine
and want to restrict the number of threads each client uses in order to
reduce heap fragmentation and CPU contention.

### Does this change impact existing behavior?

No, this only allows for overriding the default value.

### Does this change need a changelog entry? Does it require a version
change?

No, there are no breaking changes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Eric Heinz <eheinz@amazon.com>
Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>
Co-authored-by: Eric Heinz <eheinz@amazon.com>
Co-authored-by: Daniel Carl Jones <djonesoa@amazon.com>