Add expected bucket owner test for shared XZ cache (#1241)
Add the expected bucket owner test. The test attempts to use a bucket in
another AWS account as a cache and verifies that:

- write to the cache is not done if the expected bucket owner check is
enabled
- write to the cache is done if the check is disabled
- write to the cache is done if the cache bucket belongs to the expected
account

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>