Add support for CRC64-NVME checksum algorithm (#1235)
This change adds support for the CRC64-NVME checksum algorithm when
using relevant operations in `mountpoint-s3-client` and when appending
to existing objects through Mountpoint (using `--incremental-upload`
mode).

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Yes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>
Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>
Co-authored-by: Daniel Carl Jones <djonesoa@amazon.com>