Add to cache read failure log 'falling back to S3' (#1245)
When a cache read fails, there have been open questions in support cases
about if that will have then failed the whole read or asked S3. This is
a simple change to address that ambiguity.

### Does this change impact existing behavior?

Simple logging change only.

### Does this change need a changelog entry? Does it require a version
change?

No, simple log content change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>