Use custom endpoint and region setting for all fork tests (#1249)
Some of our fork tests did not set the custom endpoint for the binary,
if this is provided. This change should fix all of them to use a custom
endpoint and to set the region, if that is configured.

Does not need a changelog entry, as it only affects tests.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Christian Hagemeier <chagem@amazon.com>