Export missing types in mountpoint-s3-client API (#1248)
The S3 client API in `mountpoint-s3-client` used many types directly
from `mountpoint-s3-crt`. This change makes sure that all types used in
the API are re-exported in `mountpoint-s3-client`, so that users do not
need to depend on `mountpoint-s3-crt` directly.

### Does this change impact existing behavior?

No functional changes.

### Does this change need a changelog entry? Does it require a version
change?

Yes, it includes an entry in the `client` changelog.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>