Update semantics doc to clarify when Mountpoint can read during uploads (#1259)
To avoid ambiguity, this change updates the semantics document to be
clear that restrictions on what files can be read by Mountpoint while it
is performing a write apply to both upload modes.

### Does this change impact existing behavior?

No change in behavior.

### Does this change need a changelog entry? Does it require a version
change?

No, docs change only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>