Update reftests with small refactor and renames for clarity (#1225)
This change makes minor updates to improve clarity in the reference
tests.

### Does this change impact existing behavior?

No, refactors reftests only.

### Does this change need a changelog entry?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>