Fix fs_benchmark to correctly configure backpressure (#1260)
When running the benchmark script, it fails to run due to client errors
where backpressure isn't enabled. This is due to Mountpoint's prefetcher
relying on this being enabled, or returning
`BackpressurePreconditionFailed`.

This change configures the backpressure on the S3 client used by this
benchmark and has been tested on my own Linux machine.

### Does this change impact existing behavior?

Fixes a benchmark script only.

### Does this change need a changelog entry? Does it require a version
change?

No, benchmark script change only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>