Add timer metric over endpoint rule engine resolution (#1262)
Today, Mountpoint evaluates the rule engine with every request despite
always using the same endpoint in its default configuration.

This change adds a histogram metric so that we can begin to have
visibility on how long this is taking, and how this can be distributed.

### Does this change impact existing behavior?

No, it adds a new metric for endpoint resolution.

### Does this change need a changelog entry? Does it require a version
change?

No, metrics are not considered a stable feature. This is a new metric
addition only, no change or removal.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>