Add negative metadata cache ttl (#1246)
Adds a new CLI argument `--negative-cache-ttl` that lets you set the TTL
for negative metadata entries separately from `--metadata-ttl`. My use
case is a write once read many bucket. Objects do not get deleted from
this bucket, and new objects are added every few minutes. I'd like to be
able to set `--metadata-ttl indefinite` and `--negative-cache-ttl 60` to
effectively utilize the caching while still being able to pick up new
objects. There is an open issue for this here -
https://github.com/awslabs/mountpoint-s3/issues/831

### Does this change impact existing behavior?

No, if `--negative-cache-ttl` is omitted the existing behavior is
maintained (use `--metadata-ttl` or the default file_ttl).

### Does this change need a changelog entry? Does it require a version
change?

Because this is a new feature I believe it would require both.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: notoriaga <smeyer@fastmail.com>
Signed-off-by: Steven Meyer <smeyer@fastmail.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>