Add experimental config to write Mountpoint PID to file (#1261)
When investigating performance, we wanted to automate the collection of
profiler captures using a tool like `perf`. To do this, we needed the
process ID of Mountpoint. By writing out the PID to a file, scripts
could automatically record profiles for the lifetime of Mountpoint by
providing its PID to `perf`.

This change adds the ability to write Mountpoint's PID to a file under
an experimental/unstable environment variable. Since its unclear if we
want to expose this properly such as providing a CLI argument, we are
taking the unstable environment variable approach to make clear this
configuration may change or be removed in future.

### Does this change impact existing behavior?

This change adds a new experimental feature to write Mountpoint's PID to
a file.

### Does this change need a changelog entry? Does it require a version
change?

Since this is adding an experimental feature, no changelog entry is
required. No minor version patch is required, as this is not a stable
feature addition.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>