Increase version to 1.15 and update CHANGELOG after adding new metadata TTL flag (#1265)
The change in #1246 requires a minor version increase and a new entry in
the changelog.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>