Update caching documentation (#1267)
Updated the caching documentation to specify how the metadata cache
interacts with the data cache options.

Closes #1263.

### Does this change impact existing behavior?

No

### Does this change need a changelog entry? Does it require a version
change?

No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Simon Beal <simobeal@amazon.com>