Enforce sse on writes to the xz cache (#1257)
With this PR, flags `--sse` and `--sse-kms-key-id` will also apply to
object uploads to the xz cache.

### Does this change impact existing behavior?

Yes, bumped the version.

### Does this change need a changelog entry? Does it require a version
change?

Yes, added an entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Signed-off-by: Volodkin Vladislav <vladvolodkin@gmail.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Alessandro Passaro <alessandro.passaro@gmail.com>