Fuser fork submodule (#1269)
### Overview
To be prepared for further refactoring we want to ease the supporting
effort for the FUSER fork we're using.
As a first step we want to embed it as a git submodule which will be
using the existing `fuser/fork` branch as a remote.
This will allow us to keep the current sync flow with the original FUSER
[repo](https://github.com/cberner/fuser) but skip the manual sync step
in our `main` branch.

We will be able to update the fork by running `git submodule update
--remote mountpoint-s3-fuser` which is more idiomatic. So we still be
able to test changes locally w/o publishing the fork.

*NB* We will need to commit the changes (dirty indices) after
submodule's update.

### Does this change impact existing behavior?

This change is completely internal and does not impact customer-facing
behavior.

### Does this change need a changelog entry? Does it require a version
change?

This change does not require version or changelog changes.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Evgeny (Zhenia) Dolgii <evdolgy@amazon.com>