Add initial version of benchmark experiment runner (#1266)
In order to investigate performance in Mountpoint, we want to be able to
vary different parameters. In fact, it can be very useful to vary these
parameters together to see how performance (such as sequential read
throughput) changes as we vary two parameters together.

This change introduces a new benchmark running script which uses the
Python framework Hydra to enumerate combinations of parameters, and then
execute some function with each combination. The script manages the
lifecycle of the `mount-s3` file system and collecting data into an
output folder.

The change currently does not reuse the FIO definitions used by our
regression benchmarks. In the mid-term, these should be reconciled.

This pull request (PR) supersedes a previous PR:
https://github.com/awslabs/mountpoint-s3/pull/986.

### Does this change impact existing behavior?

No, this adds a new benchmark runner and benchmark definitions. This
does not impact the Mountpoint file system.

### Does this change need a changelog entry? Does it require a version
change?

No, no impact to Mountpoint file system or crates.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>