Add ETag to complete upload debug log (#1282)
Small change to add etag to debug logs when an MPU completes.

We already have size and object key, so this is the only missing
information.

### Does this change impact existing behavior?

Adds etag to debug logs only.

### Does this change need a changelog entry? Does it require a version
change?

No, minor logging change only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>