Add EC2 instance ID to benchmark output metadata (#1281)
We want to include the instance type in the metadata for a given
benchmark run.

This change adds a check into IMDS to query this data and add the EC2
instance type if available.

### Does this change impact existing behavior?

It adds a new field to the benchmark output metadata file.

### Does this change need a changelog entry? Does it require a version
change?

No, no change to Mountpoint itself.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>