Build and validate SLES package (#1278)
Build a separate package for SUSE Linux Enterprise Server (SLES), where
`libfuse.so.2` is delivered by `libfuse2` rpm package (as compared to
`fuse-libs` for AL2).

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

May be? Added.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Vlad Volodkin <vlaad@amazon.com>
Co-authored-by: Vlad Volodkin <vlaad@amazon.com>