Install active toolchain by default with `rustup` in `package/Dockerfile` (#1299)
Starting with v1.28, `rustup` will not install active toolchain
automatically:
https://blog.rust-lang.org/2025/03/02/Rustup-1.28.0.html#whats-new-in-rustup-1280.
This PR updates `package/Dockerfile` to install active toolchain
automatically if it's not installed.

This should fix our recent CI failures in packaging step:
https://github.com/awslabs/mountpoint-s3/actions/runs/13633623152/job/38107451366

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>