Rename Mountpoint's fuser fork (#1295)
### Overview
We renamed our fuser fork and reset its version info as we're staring to
publish it to crates.io.
In this PR we
* revert the submodule approach for consuming the fork's crate
* move the code from `vendor/fuser` to `mountpoint-s3-fuser` folder
* consume the renamed and synched FUSER Fork  from `fuser/fork`
* update dependency
* change `vendor/fuser` to `mountpoint-s3-fuser` in aux scripts and
documentation

### Does this change impact existing behavior?
This change does not impact the current behavior it only affects the
project's structure.

### Does this change need a changelog entry? Does it require a version
change?
It does not require neither version change not change log entry.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Evgeny (Zhenia) Dolgii <evdolgy@amazon.com>