Fix benchmark script column names for sweeped params (#1280)
If the column names are not defined in the Hydra script, they will be
prefixed by `+`. In order to keep things simple on the analysis side,
we'll name the columns with placeholder values which will be replaced
when running the script with multiple experiments.

### Does this change impact existing behavior?

It changes output of the benchmark script only.

### Does this change need a changelog entry? Does it require a version
change?

No, no changelog entry or version change as no change to Mountpoint
itself.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>