Remove prefix from inodes (#1303)
When Mountpoint is configured with the `--prefix` flag, all S3 requests
contain the specified prefix as part of the key. Currently, the prefix
is duplicated in each `Inode` entry in the `full_key` field. This change
remove the unnecessary duplication by only storing the partial `key` and
reconstructing the `full_key` by adding the prefix before performing any
S3 request.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

`mountpoint-s3` changelog entry. No version change.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>