Prevent failures in benchmark actions (#1307)
Benchmarks currently fail when recording a worse than 2x regression.
However, failed runs are not included in the workflow summary or in the
[performance
charts](https://github.com/awslabs/mountpoint-s3/blob/main/doc/BENCHMARKING.md).
With this change, a regression will only result in an alert, and not
lead to an action failure.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

No.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Alessandro Passaro <alexpax@amazon.co.uk>