Add network config to all component benchmarks (#1284)
This change introduces both the CRT's target network throughput
configuration and the network interface configuration to each of the
benchmarks for layers/components in Mountpoint's read path.

These are added primarily to support performance investigations, so we
can identify where there are gaps in performance and narrow them to
improve throughput of Mountpoint overall.

The target throughput default of 10.0 Gbps is removed on the lowest
level of the benchmark, given we don't know what the default is for the
CRT itself. It is left in place on all other layers as we default the
value to 10.0 Gbps inside Mountpoint's S3 client.

### Does this change impact existing behavior?

No, adds new arguments to benchmark scripts only. Even in those scripts,
we alias any command line arguments that change.

### Does this change need a changelog entry? Does it require a version
change?

No, this is benchmarking changes only.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: Daniel Carl Jones <djonesoa@amazon.com>