Pass `--foreground` in `mounthelper.go` to easily access Mountpoint logs (#1308)
Update sample `mounthelper.go` for FUSE file descriptor mounting to see
Mountpoint logs in stdout to understand what's going on easily.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Signed-off-by: Burak Varlı <burakvar@amazon.co.uk>